이벤트 지도

+++ app/src/main/java/com/example/getsumfoot/data/EventMapData.java
@@ -1,15 +1,17 @@
 package com.example.getsumfoot.data;
 
+import org.jetbrains.annotations.NotNull;
+
 public class EventMapData {
 
-    public int 위도,경도;
+    public double 위도,경도;
 
-    public EventMapData(int 위도, int 경도){
+    public EventMapData(double 위도, double 경도){
         this.위도 = 위도;
         this.경도 = 경도;
     }
 
-        public int get_latitude(){
+        public double get_latitude(){
             return 위도;
         }
 
@@ -17,13 +19,14 @@ public class EventMapData {
             this.위도 = 위도;
         }
 
-        public int get_longitude(){
+        public double get_longitude(){
             return 경도;
         }
 
         public void set_longitude(int 경도) {
             this.경도 = 경도;
         }
+
 
     }
